--- a/torchci/clickhouse_queries/test_time_per_class_periodic_jobs/query.sql
+++ b/torchci/clickhouse_queries/test_time_per_class_periodic_jobs/query.sql
@@ -2,7 +2,7 @@
 WITH good_periodic_sha AS (
     SELECT job.head_sha AS sha
     FROM
-        default.workflow_job job final
+        default.workflow_job job
     JOIN default.push ON job.head_sha = push.head_commit.'id'
     WHERE
         job.workflow_name = 'periodic'
@@ -22,30 +22,18 @@ WITH good_periodic_sha AS (
         3
 ),
 
-workflow AS (
-    SELECT id
-    FROM
-        default.workflow_run FINAL
-    WHERE
-        id IN (
-            SELECT id FROM materialized_views.workflow_run_by_head_sha w
-            WHERE head_sha IN (SELECT sha FROM good_periodic_sha)
-        )
-        AND name = 'periodic'
-),
-
 job AS (
     SELECT
+        distinct
         j.name,
         j.id,
         j.run_id
     FROM
-        default.workflow_job j FINAL
+        default.workflow_job j
     WHERE j.id IN (
         SELECT id FROM materialized_views.workflow_job_by_head_sha
         WHERE head_sha IN (SELECT sha FROM good_periodic_sha)
     )
-    AND j.run_id IN (SELECT id FROM workflow)
 ),
 
 class_duration_per_job AS (
@@ -62,7 +50,7 @@ class_duration_per_job AS (
         /* cpp tests do not populate `file` for some reason. */
         /* Exclude them as we don't include them in our slow test infra */
         test_run.file != ''
-        AND test_run.workflow_id IN (SELECT id FROM workflow)
+        AND test_run.workflow_id IN (SELECT run_id FROM job)
     GROUP BY
         test_run.invoking_file,
         test_run.classname,