--- a/torchci/clickhouse_queries/test_time_per_class_periodic_jobs/query.sql
+++ b/torchci/clickhouse_queries/test_time_per_class_periodic_jobs/query.sql
@@ -23,8 +23,7 @@ WITH good_periodic_sha AS (
 ),
 
 job AS (
-    SELECT
-        distinct
+    SELECT DISTINCT
         j.name,
         j.id,
         j.run_id