--- a/torchci/clickhouse_queries/test_time_per_file/params.json
+++ b/torchci/clickhouse_queries/test_time_per_file/params.json
@@ -1,4 +1,4 @@
 {
   "params": {},
-  "tests": []
+  "tests": [{}]
 }