--- a/torchci/clickhouse_queries/test_time_per_file_periodic_jobs/params.json
+++ b/torchci/clickhouse_queries/test_time_per_file_periodic_jobs/params.json
@@ -1,4 +1,4 @@
 {
   "params": {},
-  "tests": []
+  "tests": [{}]
 }